fix: error after changed table mode in overview
