+++ web/src/pages/Platform/Overview/components/CommonTable/index.less
